parkstay: refactor maximum stay history and make it into its own component
